requirements.txt
numpy==1.26.4
opencv_python==4.6.0.66
opencv_python_headless==4.10.0.84
PyYAML==6.0.1
torch==1.12.0+cu113
ultralytics==8.2.23
~orch==1.12.0+cu113
